Compile fixes for gtk-1.2
Original commit message from CVS:
Compile fixes for gtk-1.2

+++ sys/v4l/gstv4lsrc.h
@@ -74,7 +74,7 @@ struct _GstV4lSrc {
   /* caching values */
   gint width;
   gint height;
-  guint16 palette;
+  gint palette;
 };
 
 struct _GstV4lSrcClass {
